Extract topic store query
