+++ packages/column-live-view-engine/src/topic-store.ts
@@ -1,20 +1,5 @@
 import { Effect } from "effect";
-import {
-  acquireMaterializedQueryExecution,
-  acquireRawQueryExecution,
-  activeStoreRawQueryExecutionCount,
-  evaluateRawQuery,
-  releaseMaterializedQueryExecution,
-  releaseRawQueryExecution,
-} from "./active-query";
-import {
-  evaluateCompiledGroupedQuery,
-  prepareGroupedQuery,
-  type CompiledGroupedQuery,
-} from "./grouped-query-compiler";
-import { makeIncrementalGroupedQueryExecution } from "./grouped-incremental-execution";
-import { prepareRawQuery, type CompiledRawQuery } from "./raw-query-compiler";
-import type { QueryEvaluation } from "./query-result";
+import { activeStoreRawQueryExecutionCount } from "./active-query";
 import { collectTopicStoreHealthView, type TopicStoreHealthState } from "./topic-store-health";
 import {
   TopicStore,
@@ -22,8 +7,6 @@ import {
   topicStoreReadModel,
   topicStoreState,
 } from "./topic-store-state";
-
-type RowObject = object;
 
 export { TopicStore, topicStoreRawQueryMetadata, topicStoreReadModel };
 export {
@@ -40,75 +23,17 @@ export {
   registerTopicStoreSubscription,
   trackTopicStoreSubscriptionQueueDepth,
 } from "./topic-store-subscription";
+export {
+  acquireTopicStoreMaterializedQueryExecution,
+  acquireTopicStoreRawQueryExecution,
+  evaluateTopicStoreGroupedQuery,
+  evaluateTopicStoreRawQuery,
+  prepareTopicStoreGroupedQuery,
+  prepareTopicStoreRawQuery,
+  releaseTopicStoreMaterializedQueryExecution,
+  releaseTopicStoreRawQueryExecution,
+} from "./topic-store-query";
 export type { TopicStoreSubscriptionPermit } from "./topic-store-state";
-
-export const prepareTopicStoreRawQuery = Effect.fn(
-  "ColumnLiveViewEngine.topicStore.query.raw.prepare",
-)(function* <ResultRow extends RowObject>(store: TopicStore, query: unknown) {
-  return yield* prepareRawQuery<object, ResultRow>(
-    store.topic,
-    topicStoreState(store).storage.rawQueryMetadata,
-    query,
-  );
-});
-
-export const prepareTopicStoreGroupedQuery = Effect.fn(
-  "ColumnLiveViewEngine.topicStore.query.grouped.prepare",
-)(function* <ResultRow extends RowObject>(store: TopicStore, query: unknown) {
-  return yield* prepareGroupedQuery<object, ResultRow>(
-    store.topic,
-    topicStoreState(store).storage.rawQueryMetadata,
-    query,
-  );
-});
-
-export const evaluateTopicStoreRawQuery = <ResultRow extends RowObject>(
-  store: TopicStore,
-  compiled: CompiledRawQuery<object, ResultRow>,
-): QueryEvaluation<ResultRow> =>
-  evaluateRawQuery(topicStoreState(store).storage.readModel, compiled);
-
-export const evaluateTopicStoreGroupedQuery = <ResultRow extends RowObject>(
-  store: TopicStore,
-  compiled: CompiledGroupedQuery<object, ResultRow>,
-): QueryEvaluation<ResultRow> =>
-  evaluateCompiledGroupedQuery(topicStoreState(store).storage.readModel, compiled);
-
-export const acquireTopicStoreRawQueryExecution = Effect.fn(
-  "ColumnLiveViewEngine.topicStore.query.raw.acquire",
-)(function* <ResultRow extends RowObject>(
-  store: TopicStore,
-  compiled: CompiledRawQuery<object, ResultRow>,
-) {
-  return yield* acquireRawQueryExecution(topicStoreState(store).storage.readModel, compiled);
-});
-
-export const releaseTopicStoreRawQueryExecution = <ResultRow extends RowObject>(
-  store: TopicStore,
-  compiled: CompiledRawQuery<object, ResultRow>,
-): Effect.Effect<void> =>
-  releaseRawQueryExecution(topicStoreState(store).storage.readModel, compiled);
-
-export const acquireTopicStoreMaterializedQueryExecution = Effect.fn(
-  "ColumnLiveViewEngine.topicStore.query.materialized.acquire",
-)(function* <ResultRow extends RowObject>(
-  store: TopicStore,
-  compiled: CompiledGroupedQuery<object, ResultRow>,
-) {
-  const readModel = topicStoreState(store).storage.readModel;
-  return yield* acquireMaterializedQueryExecution(
-    readModel,
-    compiled.cacheKey,
-    (releaseRetainedChanges) =>
-      makeIncrementalGroupedQueryExecution(readModel, compiled, releaseRetainedChanges),
-  );
-});
-
-export const releaseTopicStoreMaterializedQueryExecution = <ResultRow extends RowObject>(
-  store: TopicStore,
-  compiled: CompiledGroupedQuery<object, ResultRow>,
-): Effect.Effect<void> =>
-  releaseMaterializedQueryExecution(topicStoreState(store).storage.readModel, compiled.cacheKey);
 
 const topicStoreHealthState = (store: TopicStore, activeViews: number): TopicStoreHealthState => {
   const state = topicStoreState(store);
